Fixed serious bugs with the calculation and plot of instance importance scores
I had forgotten to deactivate dropout, so it was giving stochastic outputs, the sorting method was messing up the scores and the colorscale wasn't working properly. Now the priority is to find a way to make the colors stand out more, perhaps by applying a non-linear function to the scores. Besides that, I could also account for the variation brought into the output, comparing to the previous one, through a weighted average, as well as adapt the plot's zoom so that it always has the same scale.