Fixed an issue with the denormalization method and finally reached the first satisfactory version of the instance importance scores
Denormalization was changing a variable that I didn't want to change, apparently passing it by reference instead of by value. In regards to the instance importance scores, I've decided to go forward with the weighted average formula, with 70% given to occlusion and 30% to instant output variation, followed by a tanh function with 4x multiplier. Now the scores make more sense, counting with the variation that each instance bring to the output, and they are in a more reasonable scale (closer to magnitude 1).