Started working on mask filters, created a separate script to build model interpreters and fixed issues with parameter optimization
Also did some more analysis on the SHAP values, which leads me to believe that, in the current framework, there's no way of avoiding the interference of padding values in the importance scores. As such, I'll have to use a different Shapley values package or build my own, if I want to keep using Shapley values.